add ImageStimulus, VideoStimulus
